+++ src/main/java/ua/com/javarush/tchaban/island_app/basicitem/animals/carnivorous/Carnivorous.java
@@ -1,0 +1,9 @@
+package ua.com.javarush.tchaban.island_app.basicitem.animals.carnivorous;
+
+import lombok.NoArgsConstructor;
+import ua.com.javarush.tchaban.island_app.basicitem.animals.Animal;
+
+@NoArgsConstructor
+public abstract class Carnivorous extends Animal {
+
+}
