+++ src/main/java/ua/com/javarush/tchaban/island_app/constants/ConstantsPlants.java
