+++ src/main/java/ua/com/javarush/tchaban/island_app/island/Position.java
