--- src/ua/com/javarush/tchaban/island_app/actions/PlantAbilities.java
@@ -1,5 +1,0 @@
-package ua.com.javarush.tchaban.island_app.actions;
-
-public interface PlantAbilities {
-    void regeneration();
-}
